requirements.txt
flask
flask-SQLAlchemy
# marshmallow
flask-Migrate
flask-restful
python-dotenv
pytest
requests
psycopg2
pytest-html